requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
dimod==0.8.18
dwave-cloud-client==0.6.2
dwave-hybrid==0.4.1
dwave-neal==0.5.1
dwave-networkx==0.8.3
dwave-ocean-sdk==1.5.0
dwave-qbsolv==0.2.10
dwave-system==0.8.0
dwave-tabu==0.2.1
dwavebinarycsp==0.0.12
homebase==1.0.1
idna==2.8
minorminer==0.1.9
networkx==2.4
numpy==1.15.4
ortools==6.7.4973
penaltymodel==0.16.2
penaltymodel-cache==0.4.0
penaltymodel-lp==0.1.0
penaltymodel-mip==0.2.1
plucky==0.4.3
protobuf==3.11.1
pyqubo==0.4.0
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.22.0
scipy==1.3.3
six==1.13.0
urllib3==1.25.7
